requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
cryptography==46.0.5
distlib==0.4.0
distro==1.9.0
filelock==3.24.0
google==3.0.0
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
jiter==0.13.0
nodeenv==1.10.0
platformdirs==4.9.1
pre_commit==4.5.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruff==0.15.1
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
websockets==15.0.1
